# PlayIterator, refactored