--- a/lib/ansible/executor/play_iterator.py
+++ b/lib/ansible/executor/play_iterator.py
@@ -71,7 +71,6 @@ class HostState:
         self.update_handlers = True
         self.pending_setup = False
         self.child_state = None
-        self.did_rescue = False
         self.did_start_at_task = False
 
     def __repr__(self):
@@ -80,7 +79,7 @@ class HostState:
     def __str__(self):
         return ("HOST STATE: block=%d, task=%d, rescue=%d, always=%d, handlers=%d, run_state=%s, fail_state=%s, "
                 "pre_flushing_run_state=%s, update_handlers=%s, pending_setup=%s, "
-                "child state? (%s) did rescue? %s, did start at task? %s" % (
+                "child state? (%s), did start at task? %s" % (
                     self.cur_block,
                     self.cur_regular_task,
                     self.cur_rescue_task,
@@ -92,7 +91,6 @@ class HostState:
                     self.update_handlers,
                     self.pending_setup,
                     self.child_state,
-                    self.did_rescue,
                     self.did_start_at_task,
                 ))
 
@@ -126,7 +124,6 @@ class HostState:
         new_state.pre_flushing_run_state = self.pre_flushing_run_state
         new_state.update_handlers = self.update_handlers
         new_state.pending_setup = self.pending_setup
-        new_state.did_rescue = self.did_rescue
         new_state.did_start_at_task = self.did_start_at_task
         if self.child_state is not None:
             new_state.child_state = self.child_state.copy()
@@ -365,7 +362,6 @@ class PlayIterator:
                         if len(block.rescue) > 0:
                             state.fail_state = FailedStates.NONE
                         state.run_state = IteratingStates.ALWAYS
-                        state.did_rescue = True
                     else:
                         task = block.rescue[state.cur_rescue_task]
                         if isinstance(task, Block):
@@ -399,7 +395,6 @@ class PlayIterator:
                             state.cur_always_task = 0
                             state.run_state = IteratingStates.TASKS
                             state.child_state = None
-                            state.did_rescue = False
                     else:
                         task = block.always[state.cur_always_task]
                         if isinstance(task, Block):
@@ -489,24 +484,7 @@ class PlayIterator:
     def _check_failed_state(self, state):
         if state is None:
             return False
-        elif state.run_state == IteratingStates.RESCUE and self._check_failed_state(state.child_state):
-            return True
-        elif state.run_state == IteratingStates.ALWAYS and self._check_failed_state(state.child_state):
-            return True
-        elif state.fail_state != FailedStates.NONE:
-            if state.run_state == IteratingStates.RESCUE and state.fail_state & FailedStates.RESCUE == 0:
-                return False
-            elif state.run_state == IteratingStates.ALWAYS and state.fail_state & FailedStates.ALWAYS == 0:
-                return False
-            else:
-                return not (state.did_rescue and state.fail_state & FailedStates.ALWAYS == 0)
-        elif state.run_state == IteratingStates.TASKS and self._check_failed_state(state.child_state):
-            cur_block = state._blocks[state.cur_block]
-            if len(cur_block.rescue) > 0 and state.fail_state & FailedStates.RESCUE == 0:
-                return False
-            else:
-                return True
-        return False
+        return state.fail_state != FailedStates.NONE and state.run_state == IteratingStates.COMPLETE
 
     def is_failed(self, host):
         s = self.get_host_state(host)