… lib/ansible/plugins/strategy/__init__.py
… test/units/executor/test_play_iterator.py