##### SUMMARY
This is a series of refactorings split into separate, fairly simple, commits for easier review. The benefits are:
* Less code, ~100 lines of `PlayIterator` code is being removed.
* Easier to reason about and debug.
* Performance benefits. See the Additional information below, those are isolated tests, some of them with ridiculous number of nested blocks. I am not sure how noticeable it will be in playbooks in general but I imagine it should save some time in those with a lot of includes. I have not done memory profiling but with reduced `HostState` copying this PR introduces it might be interesting to see as well.
* Changes are made only to internals of `PlayIterator` and should not be visible to its users.

There are more things to improve, I have mentioned some of them in FIXMEs/TODOs but this isn't small as-is so leaving them for next time.

##### ISSUE TYPE

<!--- Pick one below and delete the rest -->

- Feature Pull Request

##### ADDITIONAL INFORMATION
Iterating through `N` nested blocks, each block has one task and a nested block, the most nested block is being failed.

_Before:_
0.00024080276489257812 seconds, 10 blocks
0.004384040832519531 seconds, 50 blocks
0.9080429077148438 seconds, 350 blocks
**19.866453170776367** seconds, 992 blocks (I believe 992 was a limit for `devel` to handle)

_After:_
0.00009512901306152344 seconds, 10 blocks
0.0005481243133544922 seconds, 50 blocks
0.01589512825012207 seconds, 350 blocks
**0.11786723136901855** seconds, 992 blocks